Options and Parameters
Added additional asserts to the device_options testing as well as
added a test_device_paramaters unit test. These are pulling from the
mock options for the moment which seems to be working well.

Additionally added a function to set device options. I don't know that
it needs to be tested as it is largely a pass through to the SANE
device.