Set device option
Added unit tests and functionality to set device options for a device
within desanity. There is still a unit test that still fails but will
require a significant amount of mocking to get working. I'm keeping
the unit test is failing as well as commented out attempts at mocking
the MockBrother object for the setattr. It works in practice and not
in mock which isn't great but also is a bit weird at mocking.